+++ CHANGELOG.md
@@ -30,11 +30,15 @@
 
 ### 🐛 Bug Fixes
 
+- *(flake)* Turn off trufflehog checks for darwin - ([67547df](https://github.com/JayRovacsek/vulnix-pre-commit/commit/67547df0e06090401143825e06dd7c4ada1233f8))
+
 - *(flake)* Correct issue where merge clobbers pre-commit checks - ([0ad22ba](https://github.com/JayRovacsek/vulnix-pre-commit/commit/0ad22baa54b6f8f30fdcc066eada58766d4cd1cc))
 
 - *(flake)* Add standard-outputs to flake to resolve github workflow issue - ([8db642e](https://github.com/JayRovacsek/vulnix-pre-commit/commit/8db642e5d83f9f2c991cd29abbd6ac61941b3b5b))
 
 - *(flake)* Resolve duplicated inputs - ([aa3f791](https://github.com/JayRovacsek/vulnix-pre-commit/commit/aa3f7914cd009ec4ce530e1e3d7c2d20d3a50891))
+
+- *(package)* Resolve issue with typos - ([67547df](https://github.com/JayRovacsek/vulnix-pre-commit/commit/67547df0e06090401143825e06dd7c4ada1233f8))
 
 - *(package)* Resolve github checks value for pull_request - ([aa3f791](https://github.com/JayRovacsek/vulnix-pre-commit/commit/aa3f7914cd009ec4ce530e1e3d7c2d20d3a50891))